Showed archived movies table after admin archives movies, updated Archived Movies schema to match Movies schema

+++ initialization/dejaView.sql
@@ -95,7 +95,7 @@ CREATE TABLE ARCHIVED_MOVIES
  actors VARCHAR(800),
  director VARCHAR(50),
  duration INT,
- rating DOUBLE(2,1), 
+ rating DOUBLE(4,1), 
  releaseYear INT,
  updated_on DATE,
  CHECK (rating <= 10 and rating > 0),
